eliminate static cast

+++ source/tessellation/Delaunay.cpp
@@ -1641,10 +1641,10 @@ pair<vector<vector<int> >, vector<int> > Delaunay::BuildBoundary
 	return FindOuterPoints2(tproc,edges,to_duplicate.first, to_duplicate.second,box_edges, Nghost);
 }
 
-int Delaunay::GetOrgIndex(size_t index)const
+size_t Delaunay::GetOrgIndex(size_t index)const
 {
 	if (index < olength)
-		return static_cast<int>(olength);
+		return olength;
 	else
 		return OrgIndex.at(index - 3 - static_cast<int>(olength));
 }
